requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
requests 